Style: initListTotalLengh 파일 분리
