+++ src/js/events/inputBar/index.js
@@ -5,6 +5,8 @@ import { selectPaymentList } from "./selectPaymentList.js";
 import { clickPaymentHead } from "./clickPaymentHead.js";
 import { selectCategoryList } from "./selectCategoryList.js";
 import { clickCategoryHead } from "./clickCategoryHead.js";
+import { activateEditBtn } from "./activateEditBtn.js";
+import { clickEditBtn } from "./clickEditBtn.js";
 
 export const inputBarEventHandler = () => {
   $("#price-input").addEventListener("keyup", priceInput);
@@ -12,8 +14,12 @@ export const inputBarEventHandler = () => {
   $(".payment-select-head").addEventListener("click", clickPaymentHead);
   $(".payment-list-member").addEventListener("click", selectPaymentList);
  
-  document.querySelector(".category-list-member").addEventListener("click", selectCategoryList);
-  document.querySelector(".category-income-member").addEventListener("click", selectCategoryList);
-  document.querySelector(".category-select-head").addEventListener("click", clickCategoryHead);
+  // $(".category-select-head").addEventListener("click", clickCategoryHead);
+  // $(".category-list-member").addEventListener("click", selectCategoryList);
+  // $(".category-income-member").addEventListener("click", selectCategoryList);
+
+
+  $(".input-bar-wrap").addEventListener("click", activateEditBtn);
+  $("#edit-btn").addEventListener("click",clickEditBtn)
   // editbtn클릭시/ 렌더
 };
